requirements.txt
streamlit
matplotlib
scikit-learn
numpy
Pillow